+++ source/descriptor_heap.h
@@ -18,9 +18,9 @@ namespace gfx {
         ComPtr<ID3D12DescriptorHeap> heap = nullptr;
 
     private:
-        uint64_t m_alloc_index = 0;
+        uint32_t m_alloc_index = 0;
         size_t m_descriptor_size = 0;
         D3D12_CPU_DESCRIPTOR_HANDLE m_start_cpu = {};
-        std::deque<uint64_t> m_available_recycled_descriptor_indices;
+        std::deque<uint32_t> m_available_recycled_descriptor_indices;
     };
 }
